% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineK.R
\name{determineK}
\alias{determineK}
\title{Decide the number of strata for Hi-C reproducibility analysis}
\usage{
determineK(hic_data, max_ns = 10, corr_threshold = 0.1)
}
\arguments{
\item{hic_data}{A data frame returned by \code{mergeHiC}, containing merged
Hi-C interaction data.}

\item{max_ns}{An integer specifying the maximum number of strata to consider.
The default is 10.}

\item{corr_threshold}{A numeric threshold for the minimum difference in
adjacent stratum-specific rank correlations. The default is 0.1.}
}
\value{
An integer giving the selected number of strata.
}
\description{
Determines the number of distance-defined strata using a data-driven procedure
based on rank correlations between Hi-C replicates across genomic distance.
}
\examples{
set.seed(1)    # uses random jitter
rep1 <- system.file("extdata", "rep1.txt", package = "SIDR")
rep2 <- system.file("extdata", "rep2.txt", package = "SIDR")

# Merge Hi-C replicates
hic_data <- mergeHiC(rep1, rep2, chrI = "chr18", chrJ = "chr18")

# Decide the number of strata
K <- determineK(hic_data, max_ns = 10, corr_threshold = 0.1)

}
