% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.IDR.gmcm.R
\name{get.IDR}
\alias{get.IDR}
\title{Compute global IDR}
\usage{
get.IDR(idr)
}
\arguments{
\item{idr}{A numeric vector of local IDR values.}
}
\value{
A numeric vector of global IDR values for all observations.
}
\description{
Computes the global irreproducible discovery rate (IDR) for all
observations based on local IDR values.
}
