% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pseudo.R
\name{get.pseudo.mix}
\alias{get.pseudo.mix}
\title{The inverse of marginal distribution by package idr}
\usage{
get.pseudo.mix(z, mixp, mu, sigma)
}
\arguments{
\item{z}{A numeric vector of CDF values (between 0 and 1).}

\item{mixp}{Proportion of the reproducible component.}

\item{mu}{Mean of the reproducible component.}

\item{sigma}{Standard deviation of the reproducible component.}
}
\description{
This function computes pseudo-observations corresponding to given
CDF values (\code{z}) under a two-component Gaussian mixture model.
It is adapted from the \pkg{idr} package (Li et al., 2011).
}
\references{
Li, Q., Brown, J. B., Huang, H., & Bickel, P. J. (2011).
Measuring reproducibility of high-throughput experiments.
\emph{The Annals of Applied Statistics}, 5(3), 1752–1779.
See also the \pkg{idr} package: \url{https://cran.r-project.org/package=idr}
}
