% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmcmLik.R
\name{gmcmLik}
\alias{gmcmLik}
\title{Compute log-likelihood of the SIDR model}
\usage{
gmcmLik(y1, y2, ns, ind, thets)
}
\arguments{
\item{y1}{A numeric vector of Hi-C interaction scores
(-log p-values) from replicate 1.}

\item{y2}{A numeric vector of the same length as \code{y1} from
replicate 2.}

\item{ns}{An integer specifying the number of strata.}

\item{ind}{A list of length \code{ns}, where each element contains
the row indices corresponding to that stratum.}

\item{thets}{A numeric vector of length \code{3 * ns + 2}.
\enumerate{
\item The first \code{ns} elements: mixture proportions of the reproducible
component (\eqn{\pi_k});
\item The next \code{ns} elements: means of the reproducible
component (\eqn{\mu_k});
\item The next element: standard deviation of the reproducible
component. (\eqn{\sigma});
\item The next element: correlation coefficient of the irreproducible
component (\eqn{\rho});
\item The final \code{ns} elements: correlation scaling parameters of the
reproducible component(\eqn{\omega_k}).
}}
}
\value{
A single numeric value specifying the total log-likelihood
of all observations.
}
\description{
This function computes log-likelihood of all Hi-C observations under the
stratified Gaussian copula mixture model. It serves as the objective function
optimized in \code{\link{fit_IDR_stratified}} to estimate model parameters.
}
\seealso{
\code{\link{fit_IDR_stratified}} for parameter estimation
and \code{\link{get.idr.gmcm}} for local IDR computation.
}
