% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHiC.R
\name{mergeHiC}
\alias{mergeHiC}
\title{Merge two Hi-C replicates}
\usage{
mergeHiC(hic_rep1, hic_rep2, chrI, chrJ)
}
\arguments{
\item{hic_rep1}{Path to the text file (CSV, TSV, or TXT) of the first Hi-C
replicate. The file should be a data frame with the following columns:
\tabular{rll}{
column 1: \tab \code{chromosomeI} \tab character; chromosome name of
fragment I. \cr
column 2: \tab \code{fragmentI} \tab integer; coordinate of the start
point of fragment I. \cr
column 3: \tab \code{chromosomeJ} \tab character; chromosome name of
fragment J. \cr
column 4: \tab \code{fragmentJ} \tab integer; coordinate of the start
point of fragment J. \cr
column 5: \tab \code{pvalue} \tab numeric; p-value measuring the
significance of the interaction between the two fragments. \cr
}}

\item{hic_rep2}{Path to the text file (CSV, TSV, or TXT) of the second Hi-C
replicate. The file should have the same columns as \code{hic_rep1}.}

\item{chrI}{A character string specifying the chromosome name of fragment I
to be analyzed.}

\item{chrJ}{A character string specifying the chromosome name of fragment J
to be analyzed.}
}
\value{
A data frame containing the following columns:
\tabular{rll}{
column 1: \tab \code{chromosomeI} \tab character; chromosome name of
fragment I. \cr
column 2: \tab \code{chromosomeJ} \tab character; chromosome name of
fragment J. \cr
column 3: \tab \code{fragmentI} \tab integer; coordinate of the start
point of fragment I. \cr
column 4: \tab \code{fragmentJ} \tab integer; coordinate of the start
point of fragment J. \cr
column 5: \tab \code{pvalue1} \tab numeric; p-value for the first
replicate. \cr
column 6: \tab \code{pvalue2} \tab numeric; p-value for the second
replicate. \cr
column 7: \tab \code{obs1} \tab numeric; -log(\code{pvalue1}). \cr
column 8: \tab \code{obs2} \tab numeric; -log(\code{pvalue2}). \cr
column 9: \tab \code{dist} \tab numeric; genomic distance between
fragment I and fragment J (kb). \cr
}
}
\description{
This function pre-processes two Hi-C replicate files of the same resolution,
filters them for specified chromosomes, and merges them into a
single data frame.
}
\examples{
set.seed(1)    # uses random jitter
rep1 <- system.file("extdata", "rep1.txt", package = "SIDR")
rep2 <- system.file("extdata", "rep2.txt", package = "SIDR")

# Merge Hi-C replicates
hic_data <- mergeHiC(rep1, rep2, chrI = "chr18", chrJ = "chr18")

}
