% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyData.R
\name{stratifyData}
\alias{stratifyData}
\title{Stratify Hi-C data into distance-based strata}
\usage{
stratifyData(hic_data, ns)
}
\arguments{
\item{hic_data}{A data frame returned by \code{mergeHiC}, containing merged
Hi-C interaction data.}

\item{ns}{An integer specifying the number of strata into which Hi-C
interactions are stratified (must be >= 2).}
}
\value{
A list of length \code{ns}, where each element contains the row
indices of that stratum.
}
\description{
Divides Hi-C interactions into distance-based strata to
account for distance-dependent changes in reproducibility.
It returns a list of row indices for each stratum.
}
\examples{
set.seed(1)    # uses random jitter
rep1 <- system.file("extdata", "rep1.txt", package = "SIDR")
rep2 <- system.file("extdata", "rep2.txt", package = "SIDR")

# Merge Hi-C replicates
hic_data <- mergeHiC(rep1, rep2, chrI = "chr18", chrJ = "chr18")

# Decide the number of strata
K <- determineK(hic_data, max_ns = 10, corr_threshold = 0.1)

# Stratify data
ind <- stratifyData(hic_data, ns = K)

}
