% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePara.R
\name{choosePara}
\alias{choosePara}
\title{Decide initial parameters for Hi-C reproducibility analysis}
\usage{
choosePara(hic_data, ns, ind, prp = 0.7, rho = 0.15)
}
\arguments{
\item{hic_data}{A data frame returned by \code{mergeHiC}, containing merged
Hi-C interaction data.}

\item{ns}{An integer specifying the number of strata.}

\item{ind}{A list returned by \code{stratifyData}, containing row indices
of each stratum.}

\item{prp}{A numeric value between 0 and 1 specifying the proportion
threshold for selecting top-ranked observations (highly likely to be
reproducible) to compute correlation.}

\item{rho}{A numeric value between 0 and 1 specifying the correlation for the
bottom-ranked signals (highly likely to be irreproducible).}
}
\value{
A list containing:
\describe{
\item{mixps}{A numeric vector of reproducible proportions.}
\item{mus}{A numeric vector of means for the reproducible component in each stratum.}
\item{rho}{Correlation coefficient for the irreproducible component.}
\item{sigma}{Standard deviation (set to 1) for the reproducible component.}
\item{omega}{A numeric vector of correlation scale parameters for the reproducible component.}
}
}
\description{
Compute initial parameters for the Gaussian mixture copula model used
in stratified IDR analysis.
}
\examples{
set.seed(1)    # uses random jitter
rep1 <- system.file("extdata", "rep1.txt", package = "SIDR")
rep2 <- system.file("extdata", "rep2.txt", package = "SIDR")

# Merge Hi-C replicates
hic_data <- mergeHiC(rep1, rep2, chrI = "chr18", chrJ = "chr18")

# Decide the number of strata
K <- determineK(hic_data, max_ns = 10, corr_threshold = 0.1)

# Stratify data
ind <- stratifyData(hic_data, ns = K)

# Select initial parameters
init <- choosePara(hic_data, ns = K, ind = ind, prp = 0.7, rho = 0.15)

}
