% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IDR_stratified.R
\name{fit_IDR_stratified}
\alias{fit_IDR_stratified}
\title{Fit the SIDR Model}
\usage{
fit_IDR_stratified(hic_data, ns, ind, thet.ini)
}
\arguments{
\item{hic_data}{A data frame returned by \code{mergeHiC}, containing merged
Hi-C interaction data. Alternatively, it can be a data frame
containing well-paired observations for the two replicates (obs1 and obs2).}

\item{ns}{An integer specifying the number of strata.}

\item{ind}{A list returned by \code{stratifyData}, where each element contains
the indices of the observations belonging to that stratum.}

\item{thet.ini}{A numeric vector of length \code{3 * ns + 2}.
\enumerate{
\item The first \code{ns} elements: logit-transformed initial mixture
proportions of the reproducible component (\eqn{\pi_k});
\item The next \code{ns} elements: initial means of the reproducible
component (\eqn{\mu_k});
\item The next element: initial standard deviation of the reproducible
component (\eqn{\sigma});
\item The next element: logit-transformed initial correlation coefficient
of the irreproducible component (\eqn{\rho});
\item The final \code{ns} elements: initial correlation scaling parameters
of the reproducible component (\eqn{\omega_k}).
}}
}
\value{
A list containing:
\describe{
\item{est}{A numeric vector of estimated model parameters.}
\item{idr}{A list of local irreproducible discovery rates (idr) per stratum.}
\item{IDR}{A numeric vector of global irreproducible discovery rate (IDR).}
}
}
\description{
This function estimates parameters of the stratified Gaussian copula mixture
model using the Nelder–Mead optimization method, and computes the local and
global irreproducible discovery rate (idr and IDR) for Hi-C interactions
based on the estimated parameters. The computation is inspired by
the IDR methodology implemented in the
\href{https://bioconductor.org/packages/release/bioc/html/idr.html}{IDR package}
(Li et al., 2011).
}
\examples{
\dontrun{
# Merge Hi-C replicates
data <- mergeHiC("rep1.txt", "rep2.txt", chrI = "chr18", chrJ = "chr18")

# Stratify data
ind <- stratifyData(data, ns = 3)

# Decide initial parameters
init <- choosePara(data, ns = 3, ind = ind)

# Fit the stratified IDR model
thet.ini <- c(logit(init$mixps), init$mus, init$sigma, logit(init$rho), init$omega)
res <- fit_IDR_stratified(data, ns = 3, ind = ind, thet.ini = thet.ini)
}

}
\references{
Li, Q., Brown, J. B., Huang, H., & Bickel, P. J. (2011).
Measuring reproducibility of high-throughput experiments.
\emph{Annals of Applied Statistics}, 5(3), 1752–1779.
\url{https://bioconductor.org/packages/release/bioc/html/idr.html}
}
\seealso{
\code{\link{choosePara}} for initializing parameters, and
\code{\link{stratifyData}} for defining distance-based strata.
}
